Fix aarch64 --no-apply-dynamic-relocs option
This option is intended to make gold not apply link-time values for
absolution relocations which have dynamic relocations emitted for
them, in order to workaround an android dynamic loader bug in old
versions of android.

Unfortunately, it also had the side-effect of breaking debug data,
because the dynamic relocations are not used for non-ALLOC sections,
but the flag was also suppressing the static relocation.

This fix was proposed by jyknight@ and it fixed the problem by filtering
out non-ALLOC sections with the --no-apply-dynamic-relocs option.

BUG: 70838247
TEST: Debug info appears with --no-apply-dynamic-relocs
Signed-off-by: mydongistiny <jaysonedson@gmail.com>